Exclude answer-functions and move to another file
